MueLu: limit numCols of P used in MapTransferFact
P might contain entries for rotational nullspace vectors. For dual
variables in saddle point problems, coarsening of the interface map
should only be based on translational dofs, since rotational dual
variables are never defined.

This change introduces a parameter to the MapTransferFactory, that
allows to limit the selection of columns to be used in the map transfer
to translational modes.

Also, a test for 3D elasticity is added.