MueLu: compute region interface nodes by hand
Extract nodesets from an exodus mesh and manually compute the
intersection of nodesets (one nodesets = one region) to obtain the list
of nodes on each region interface.

Note: this is a clumsy implementation, that should be replaced by a more
elegant use of STK's Selector mechanism to perform intersection of
element blocks/nodesets.