MueLu: compute region interface sent/receive lists
At region interfaces, nodes need to be duplicated. The duplicates need
to be communicated to the neighboring ranks.

Therefore, each MPI rank needs to precompute, which LIDs and
GIDs are sent to another rank. Also, each rank needs to precompute which
LIDs and GIDs it will receive from another rank.

This has been verifired with a small 5-region mesh of a cylinder.