MueLu: concistently use LIDs/GIDs starting from 0
There has been confusion about LIDs and especially GIDs extracted from
STK. We now have concluded, that we need to ask the "mesh buld data" for
an identifier() to obtain a valid GID (starting with 1) and then
subtract 1 to start counting from 0 (Tpetra/Panzer style). To be
consistent, we also ask the "mesh bulk data" for the owning rank of any
STK entity.

This commit updates the identification of GIDs on region interfaces to
use the corrected Tpetra-sytle numbered GIDs.